requirements.txt
fastapi
uvicorn
websockets
twilio
openai
python-dotenv
deepgram-sdk
httpx
